[12.10-18:01] [DEV] added strsig function

+++ sources/executor/executor_bonus.c
@@ -88,6 +88,6 @@ void			executor(const t_cmd *command)
 			free(bin_dir);
 			free(binary_full_path);
 		}
-		ms_setenv_int("?", (int)WIFEXITED(ret_value), TRUE, TRUE); // ATTENTION !! CHANGE TRUE, TRUE to TRUE, FALSE
+		ms_setenv_int("?", WIFEXITED(ret_value), TRUE, TRUE); // ATTENTION !! CHANGE TRUE, TRUE to TRUE, FALSE
 	}
 }
